added jsdocs and fix lint errors
