+++ components/typography/Heading.tsx
@@ -1,14 +1,33 @@
-import type { HeadingProps } from "@/types/typography/Heading";
-import { HeadingLevel, HeadingTypeStyle } from "@/types/typography/Heading";
+import React from 'react';
 import { twMerge } from 'tailwind-merge';
 
+import { HeadingLevel, HeadingTypeStyle } from '@/types/typography/Heading';
+
+export interface HeadingProps {
+  typeStyle?: HeadingTypeStyle;
+  level?: HeadingLevel;
+  textColor?: string;
+  className?: string;
+  children?: React.ReactNode;
+  id?: string;
+}
+
+/**
+ * Heading
+ * @param {HeadingTypeStyle} props.typeStyle contains the type of heading style. HeadingTypeStyle.lg is default
+ * @param {HeadingLevel} props.level contains the level of heading. HeadingLevel.h2 is by default
+ * @param {string} props.textColor contains text color for the heading. 'text-primary-800' is by default
+ * @param {string} props.className contains additional classes that should be added to the component
+ * @param {React.ReactNode} props.children contains all the child elements bounded inside component
+ * @param {string} props.id contains an id to be appended on heading
+ */
 export default function Heading({
   typeStyle = HeadingTypeStyle.lg,
   level = HeadingLevel.h2,
   textColor = 'text-primary-800',
   className = '',
   children,
-  id,
+  id
 }: HeadingProps) {
   const Tag = level ?? HeadingLevel.h2;
