+++ components/typography/TextLink.tsx
@@ -1,7 +1,22 @@
-import type { TextLinkProps } from '@/types/typography/TextLink';
-import { twMerge } from 'tailwind-merge';
 import Link from 'next/link';
+import { twMerge } from 'tailwind-merge';
 
+export interface TextLinkProps {
+  href: string;
+  className?: string;
+  target?: string;
+  children?: React.ReactNode;
+  id?: string;
+}
+
+/**
+ *
+ * @param {string} props.href contains a URL as href for a link
+ * @param {string} props.className contains additional classes that should be added to the component
+ * @param {string} props.target contains the target value for the link
+ * @param {React.ReactNode} props.children contains all the child elements bounded inside component
+ * @param {string} props.id contains an id to be appended on heading
+ */
 export default function TextLink({
   href,
   className = '',
@@ -9,16 +24,17 @@ export default function TextLink({
   children,
   id,
 }: TextLinkProps) {
+  // eslint-disable-next-line max-len
   const classNames = twMerge(`text-secondary-500 underline hover:text-gray-800 font-medium transition ease-in-out duration-300 ${className || ''}`);
 
   return (
-    <Link 
-      href={href} 
-      target={target} 
-      rel="noreferrer noopener" 
+    <Link
+      href={href}
+      target={target}
+      rel='noreferrer noopener'
       className={classNames}
       id={id}
-      data-testid="TextLink-href"
+      data-testid='TextLink-href'
     >
       <span className={classNames}>
         {children}
