+++ types/typography/Paragraph.ts
@@ -3,11 +3,3 @@ export enum ParagraphTypeStyle {
     md = 'body-md',
     sm = 'body-sm',
 }
-
-export interface ParagraphProps {
-    typeStyle?: ParagraphTypeStyle;
-    textColor?: string; 
-    fontWeight?: string;
-    className?: string;
-    children?: React.ReactNode;
-}
