--- types/typography/TextLink.ts
@@ -1,7 +1,0 @@
-export interface TextLinkProps {
-    href: string;
-    className?: string;
-    target?: string;
-    children?: React.ReactNode;
-    id?: string;
-  }
